Move remotes code to submodule
