+++ docs/make.jl
@@ -23,7 +23,6 @@ makedocs(
     clean = false,
     sitename = "Documenter.jl",
     authors = "Michael Hatherly, Morten Piibeleht, and contributors.",
-    linkcheck = !("skiplinks" in ARGS),
     pages = [
         "Home" => "index.md",
         "Manual" => Any[
@@ -47,6 +46,8 @@ makedocs(
         ],
         "contributing.md",
     ],
+    repo = Remotes.GitHub("JuliaDocs", "Documenter.jl"),
+    linkcheck = !("skiplinks" in ARGS),
     strict = !("strict=false" in ARGS),
     doctest = ("doctest=only" in ARGS) ? :only : true,
 )
